% Generated by build script
\name{ftse_russell_history}
\alias{ftse_russell_history}
\title{Fetch Russell index values (full history)}
\usage{
ftse_russell_history()
}
\arguments{
\item{code}{Russell index code (e.g. "US2000", "US1000", "US3000")}
}
\value{tibble: date (Date), index_name (character),}
\description{Fetch Russell index values (full history)}
