% Generated by build script
\name{ftse_russell_indexes}
\alias{ftse_russell_indexes}
\title{List available Russell indexes}
\usage{
ftse_russell_indexes()
}
\arguments{
\item{family}{Optional family name to filter by (e.g. "Russell U.S.")}
}
\value{tibble: family (character), index (character),}
\description{List available Russell indexes}
