% Generated by build script
\name{ftse_russell_ytd}
\alias{ftse_russell_ytd}
\title{Fetch Russell index values (year-to-date)}
\usage{
ftse_russell_ytd()
}
\arguments{
\item{code}{Russell index code (e.g. "US2000" for Russell 2000,}
}
\value{tibble: date (Date), index_name (character),}
\description{Fetch Russell index values (year-to-date)}
