% Generated by build script
\name{ftse_historic_indexes}
\alias{ftse_historic_indexes}
\title{List available FTSE historic indexes}
\usage{
ftse_historic_indexes()
}
\arguments{
\item{group}{Optional group name to filter by (e.g. "FTSE Global Equity")}
\item{currency}{Optional currency code to filter by (e.g. "USD", "GBP")}
}
\value{tibble: group (character), index_name (character),}
\description{List available FTSE historic indexes}
